// src/app/models/stats.model.ts
export interface Stats {
    last_day_average_time: string;
  last_day_total_time: string;
  last_day_count: number;
  last_month_average_time: string;
  last_month_total_time: string;
  last_month_count: number;
  all_time_average_time: string;
  all_time_total_time: string;
  all_time_count: number;
}